requirements.txt
Django==1.8.7
decorator==4.0.4
django-extensions==1.5.9
djangorestframework==3.3.1
ipython==4.0.1
ipython-genutils==0.1.0
path.py==8.1.2
pexpect==4.0.1
pickleshare==0.5
ptyprocess==0.5
simplegeneric==0.8.1
six==1.10.0
traitlets==4.0.0
